refactor(weekly): weekly_tab_page.dart 분해 (884→574줄)
큰 위젯 빌더 메소드들을 part 파일로 분리하고, 순수 함수는
lib/utils로 추출 + 단위 테스트 추가.

신설 (utils):
- lib/utils/weekly_selection.dart (~30줄)
  · weekStartMonday, weeklyKeyFor, seedFromKey
- test/utils/weekly_selection_test.dart (10 tests)

신설 (part 파일):
- lib/widgets/weekly/weekly_avatar.dart (~67줄)
  · _weeklyPersonAvatar, _weeklyAvatarFallback
- lib/widgets/weekly/weekly_list_panel.dart (~258줄)
  · _buildWeeklyListPanel, _weeklyInlineQuizButton, _weeklyPersonTitleBadge

부모 파일:
- 884 → 574줄 (-35%)
- 순수 함수 호출을 weekStartMonday()/weeklyKeyFor()/seedFromKey()로 교체
- part 'weekly/...' 디렉티브 2개 추가

flutter analyze: 0 issues
flutter test: 117 tests pass (107 + 10 신규)
